LRAC-10551 Rename
