+++ modules/apps/analytics/analytics-batch-export-import-api/bnd.bnd
@@ -3,4 +3,4 @@ Bundle-SymbolicName: com.liferay.analytics.batch.exportimport.api
 Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.analytics.batch.exportimport.client,\
-	com.liferay.analytics.batch.exportimport.helper
+	com.liferay.analytics.batch.exportimport.manager
