+++ modules/apps/analytics/analytics-batch-export-import-api/src/main/resources/com/liferay/analytics/batch/exportimport/manager/packageinfo
